+++ Builder/SwaggerResolverBuilder.php
@@ -14,11 +14,15 @@ declare(strict_types=1);
 namespace Linkin\Bundle\SwaggerResolverBundle\Builder;
 
 use EXSyst\Component\Swagger\Schema;
+use Linkin\Bundle\SwaggerResolverBundle\Enum\ParameterExtensionEnum;
 use Linkin\Bundle\SwaggerResolverBundle\Enum\ParameterTypeEnum;
 use Linkin\Bundle\SwaggerResolverBundle\Exception\UndefinedPropertyTypeException;
 use Linkin\Bundle\SwaggerResolverBundle\Normalizer\SwaggerNormalizerInterface;
 use Linkin\Bundle\SwaggerResolverBundle\Resolver\SwaggerResolver;
 use Linkin\Bundle\SwaggerResolverBundle\Validator\SwaggerValidatorInterface;
+use Symfony\Component\Validator\Exception\ValidationFailedException;
+use Symfony\Component\Validator\Mapping\ClassMetadata;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 use function in_array;
 use function is_array;
@@ -44,15 +48,26 @@ class SwaggerResolverBuilder
     private $swaggerValidators;
 
     /**
+     * @var ValidatorInterface|null
+     */
+    private $validator;
+
+    /**
      * @param SwaggerValidatorInterface[] $swaggerValidators
      * @param SwaggerNormalizerInterface[] $swaggerNormalizers
      * @param array $normalizationLocations
+     * @param ValidatorInterface|null $validator
      */
-    public function __construct(array $swaggerValidators, array $swaggerNormalizers, array $normalizationLocations)
-    {
+    public function __construct(
+        array $swaggerValidators,
+        array $swaggerNormalizers,
+        array $normalizationLocations,
+        ?ValidatorInterface $validator = null
+    ) {
         $this->normalizationLocations = $normalizationLocations;
         $this->swaggerNormalizers = $swaggerNormalizers;
         $this->swaggerValidators = $swaggerValidators;
+        $this->validator = $validator;
     }
 
     /**
@@ -91,12 +106,15 @@ class SwaggerResolverBuilder
                 throw new UndefinedPropertyTypeException($definitionName, $name, $propertyType);
             }
 
-            if (!$swaggerResolver->isRequired($name)) {
+            $isNullable = $propertySchema->getExtensions()[ParameterExtensionEnum::X_NULLABLE] ?? null;
+
+            if ($isNullable === true) {
                 $allowedTypes[] = 'null';
             }
 
             $swaggerResolver->setAllowedTypes($name, $allowedTypes);
             $swaggerResolver = $this->addNormalization($swaggerResolver, $name, $propertySchema);
+            $swaggerResolver = $this->addConstraint($swaggerResolver, $name, $definition);
 
             if (null !== $propertySchema->getDefault()) {
                 $swaggerResolver->setDefault($name, $propertySchema->getDefault());
@@ -112,6 +130,57 @@ class SwaggerResolverBuilder
         }
 
         return $swaggerResolver;
+    }
+
+    /**
+     * @param SwaggerResolver $resolver
+     * @param string $name
+     * @param Schema $definition
+     *
+     * @return SwaggerResolver
+     */
+    private function addConstraint(SwaggerResolver $resolver, string $name, Schema $definition): SwaggerResolver
+    {
+        if (!$this->validator) {
+            return $resolver;
+        }
+
+        $definitionClass = $definition->getExtensions()[ParameterExtensionEnum::X_CLASS] ?? null;
+
+        if (!$definitionClass) {
+            return $resolver;
+        }
+
+        if (!$this->validator->hasMetadataFor($definitionClass)) {
+            return $resolver;
+        }
+
+        /** @var ClassMetadata $definitionMetadata */
+        $definitionMetadata = $this->validator->getMetadataFor($definitionClass);
+        $propertyMetadataList = $definitionMetadata->getPropertyMetadata($name);
+
+        foreach ($propertyMetadataList as $propertyMetadata) {
+            if (!$propertyMetadata->getConstraints()) {
+                continue;
+            }
+
+            $resolver->addAllowedValues($name, function ($value) use ($definitionClass, $name) {
+                $violations = $this->validator
+                    ->startContext()
+                    ->atPath($name)
+                    ->validatePropertyValue($definitionClass, $name, $value)
+                    ->getViolations()
+                ;
+
+                if ($violations->count() > 0) {
+                    throw new ValidationFailedException($value, $violations);
+                }
+
+                return true;
+            });
+        }
+
+        return $resolver;
     }
 
     /**