+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 ## [Unreleased]
+### Added
+- Accept and read swagger extension `x-nullable`.
+- Accept property annotation constraints.
+- Swagger extension `x-class` has full class name of body model.
+- Accept enum class to swagger enum configuration.
+### Changed
+- Required field can be with `null` value.
+- Remove support of `symfony/*` with version `~3.4`.
 
 ## [0.4.11]
 ### Changed